+++ kifi-backend/modules/graph/app/com/keepit/graph/GraphProdModule.scala
@@ -11,7 +11,7 @@ import com.kifi.franz.{SimpleSQSClient, QueueName, SQSQueue}
 
 import com.amazonaws.auth.BasicAWSCredentials
 import com.amazonaws.regions._
-import com.keepit.graph.ingestion.GraphUpdate
+import com.keepit.graph.manager.GraphUpdate
 
 
 case class GraphProdModule() extends GraphModule with CommonProdModule {
